requirements.txt
torch==2.2.2
torchvision==0.17.2
transformers==4.30.0
faiss-cpu==1.7.4
Pillow==9.5.0
Flask==2.3.0
Flask-CORS==4.0.0
numpy==1.24.0
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.65.0 